# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NOC monitoring API for hosting company website monitoring & incident detection
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dns-resolvers:
    get:
      operationId: listDnsResolvers
      tags: [dns]
      summary: List custom and built-in DNS resolvers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsResolverList"
    post:
      operationId: addDnsResolvers
      tags: [dns]
      summary: Add one or many custom DNS resolvers (bulk text or list)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDnsResolversBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddDnsResolversResult"
  /dns-resolvers/{id}:
    delete:
      operationId: deleteDnsResolver
      tags: [dns]
      summary: Delete a custom DNS resolver
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List recent system event logs (newest first)
      parameters:
        - in: query
          name: level
          schema: { type: string, enum: [debug, info, warn, error, all] }
        - in: query
          name: category
          schema: { type: string, enum: [system, monitor, incident, api, dns, all] }
        - in: query
          name: siteId
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
  /logs/export:
    get:
      operationId: exportLogs
      tags: [logs]
      summary: Download all logs as a plain text file
      parameters:
        - in: query
          name: level
          schema: { type: string, enum: [debug, info, warn, error, all] }
        - in: query
          name: category
          schema: { type: string, enum: [system, monitor, incident, api, dns, all] }
        - in: query
          name: siteId
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100000, default: 10000 }
      responses:
        "200":
          description: Plain-text log file
          content:
            text/plain:
              schema: { type: string }
components:
  schemas:
...
    LogEntry:
      type: object
      properties:
        id: { type: integer }
        timestamp: { type: string, format: date-time }
        level: { type: string, enum: [debug, info, warn, error] }
        category: { type: string, enum: [system, monitor, incident, api, dns] }
        siteId: { type: integer, nullable: true }
        message: { type: string }
        details: { type: string, nullable: true }
      required: [id, timestamp, level, category, siteId, message, details]
    DnsResolverEntry:
      type: object
      properties:
        id: { type: integer, nullable: true, description: "null for built-in resolvers" }
        name: { type: string }
        address: { type: string }
        builtIn: { type: boolean }
      required: [id, name, address, builtIn]
    DnsResolverList:
      type: object
      properties:
        builtIn:
          type: array
          items: { $ref: "#/components/schemas/DnsResolverEntry" }
        custom:
          type: array
          items: { $ref: "#/components/schemas/DnsResolverEntry" }
      required: [builtIn, custom]
    AddDnsResolversBody:
      type: object
      properties:
        text:
          type: string
          description: Bulk text — one IP per line or comma separated. Optional "name=ip" format supported.
      required: [text]
    AddDnsResolversResult:
      type: object
      properties:
        added:
          type: array
          items: { $ref: "#/components/schemas/DnsResolverEntry" }
        skipped:
          type: array
          items:
            type: object
            properties:
              value: { type: string }
              reason: { type: string }
            required: [value, reason]
      required: [added, skipped]
...